+++ src/pretix/base/__init__.py
@@ -19,48 +19,3 @@
 # You should have received a copy of the GNU Affero General Public License along with this program.  If not, see
 # <https://www.gnu.org/licenses/>.
 #
-
-# This file is based on an earlier version of pretix which was released under the Apache License 2.0. The full text of
-# the Apache License 2.0 can be obtained at <http://www.apache.org/licenses/LICENSE-2.0>.
-#
-# This file may have since been changed and any changes are released under the terms of AGPLv3 as described above. A
-# full history of changes and contributors is available at <https://github.com/pretix/pretix>.
-#
-# This file contains Apache-licensed contributions copyrighted by: Tobias Kunze
-#
-# Unless required by applicable law or agreed to in writing, software distributed under the Apache License 2.0 is
-# distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-# License for the specific language governing permissions and limitations under the License.
-
-from django.apps import AppConfig
-
-
-class PretixBaseConfig(AppConfig):
-    name = 'pretix.base'
-    label = 'pretixbase'
-
-    def ready(self):
-        from . import exporter  # NOQA
-        from . import payment  # NOQA
-        from . import exporters  # NOQA
-        from . import invoice  # NOQA
-        from . import notifications  # NOQA
-        from . import email  # NOQA
-        from .services import auth, checkin, export, mail, tickets, cart, orderimport, orders, invoices, cleanup, update_check, quotas, notifications, vouchers  # NOQA
-        from django.conf import settings
-
-        try:
-            from .celery_app import app as celery_app  # NOQA
-        except ImportError:
-            pass
-
-        if hasattr(settings, 'RAVEN_CONFIG'):
-            from ..sentry import initialize
-            initialize()
-
-
-default_app_config = 'pretix.base.PretixBaseConfig'
-try:
-    import pretix.celery_app as celery  # NOQA
-except ImportError:
-    pass
