+++ src/pretix/base/apps.py
@@ -32,13 +32,34 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 # License for the specific language governing permissions and limitations under the License.
 
-from django.urls import URLPattern
-from django.urls.resolvers import RegexPattern
+from django.apps import AppConfig
 
 
-def event_url(route, view, name=None, require_live=True):
-    if callable(view):
-        pattern = RegexPattern(route, name=name, is_endpoint=True)
-        pattern._require_live = require_live
-        return URLPattern(pattern, view, {}, name)
-    raise TypeError('view must be a callable.')
+class PretixBaseConfig(AppConfig):
+    name = 'pretix.base'
+    label = 'pretixbase'
+
+    def ready(self):
+        from . import exporter  # NOQA
+        from . import payment  # NOQA
+        from . import exporters  # NOQA
+        from . import invoice  # NOQA
+        from . import notifications  # NOQA
+        from . import email  # NOQA
+        from .services import auth, checkin, export, mail, tickets, cart, orderimport, orders, invoices, cleanup, update_check, quotas, notifications, vouchers  # NOQA
+        from django.conf import settings
+
+        try:
+            from .celery_app import app as celery_app  # NOQA
+        except ImportError:
+            pass
+
+        if hasattr(settings, 'RAVEN_CONFIG'):
+            from ..sentry import initialize
+            initialize()
+
+
+try:
+    import pretix.celery_app as celery  # NOQA
+except ImportError:
+    pass
