+++ src/pretix/helpers/apps.py
@@ -19,3 +19,11 @@
 # You should have received a copy of the GNU Affero General Public License along with this program.  If not, see
 # <https://www.gnu.org/licenses/>.
 #
+from django.apps import AppConfig
+
+from .database import *  # noqa
+
+
+class PretixHelpersConfig(AppConfig):
+    name = 'pretix.helpers'
+    label = 'pretixhelpers'
