+++ src/pretix/multidomain/apps.py
@@ -32,13 +32,9 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 # License for the specific language governing permissions and limitations under the License.
 
-from django.urls import URLPattern
-from django.urls.resolvers import RegexPattern
+from django.apps import AppConfig
 
 
-def event_url(route, view, name=None, require_live=True):
-    if callable(view):
-        pattern = RegexPattern(route, name=name, is_endpoint=True)
-        pattern._require_live = require_live
-        return URLPattern(pattern, view, {}, name)
-    raise TypeError('view must be a callable.')
+class PretixMultidomainConfig(AppConfig):
+    name = 'pretix.multidomain'
+    label = 'pretixmultidomain'
