+++ src/pretix/plugins/manualpayment/apps.py
@@ -19,3 +19,19 @@
 # You should have received a copy of the GNU Affero General Public License along with this program.  If not, see
 # <https://www.gnu.org/licenses/>.
 #
+from django.apps import AppConfig
+from django.utils.translation import gettext_lazy as _
+
+from pretix import __version__ as version
+
+
+class ManualPaymentApp(AppConfig):
+    name = 'pretix.plugins.manualpayment'
+    verbose_name = _("Manual payment")
+
+    class PretixPluginMeta:
+        name = _("Manual payment")
+        author = _("the pretix team")
+        version = version
+        category = 'PAYMENT'
+        description = _("This plugin adds a customizable payment method for manual processing.")
