+++ src/pretix/plugins/statistics/__init__.py
@@ -19,25 +19,3 @@
 # You should have received a copy of the GNU Affero General Public License along with this program.  If not, see
 # <https://www.gnu.org/licenses/>.
 #
-from django.apps import AppConfig
-from django.utils.translation import gettext_lazy as _
-
-from pretix import __version__ as version
-
-
-class StatisticsApp(AppConfig):
-    name = 'pretix.plugins.statistics'
-    verbose_name = _("Statistics")
-
-    class PretixPluginMeta:
-        name = _("Statistics")
-        author = _("the pretix team")
-        version = version
-        category = 'FEATURE'
-        description = _("This plugin shows you various statistics.")
-
-    def ready(self):
-        from . import signals  # NOQA
-
-
-default_app_config = 'pretix.plugins.statistics.StatisticsApp'
