+++ src/pretix/plugins/ticketoutputpdf/apps.py
@@ -32,13 +32,32 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 # License for the specific language governing permissions and limitations under the License.
 
-from django.urls import URLPattern
-from django.urls.resolvers import RegexPattern
+from django.apps import AppConfig
+from django.utils.functional import cached_property
+from django.utils.translation import gettext_lazy as _
+
+from pretix import __version__ as version
 
 
-def event_url(route, view, name=None, require_live=True):
-    if callable(view):
-        pattern = RegexPattern(route, name=name, is_endpoint=True)
-        pattern._require_live = require_live
-        return URLPattern(pattern, view, {}, name)
-    raise TypeError('view must be a callable.')
+class TicketOutputPdfApp(AppConfig):
+    name = 'pretix.plugins.ticketoutputpdf'
+    verbose_name = _("PDF ticket output")
+
+    class PretixPluginMeta:
+        name = _("PDF ticket output")
+        author = _("the pretix team")
+        version = version
+        category = 'FORMAT'
+        description = _("This plugin allows you to print out tickets as PDF files")
+
+    def ready(self):
+        from . import signals  # NOQA
+
+    @cached_property
+    def compatibility_errors(self):
+        errs = []
+        try:
+            import reportlab  # NOQA
+        except ImportError:
+            errs.append("Python package 'reportlab' is not installed.")
+        return errs
